Define 'join' of attached volumes w/ instances
I think of this as a client-side join where we find the Volumes attached
to a collection of instances (selected via another view). That might not
make any sense. If so, I offer you this commit message ...

We need to know what instances in the "selectedResources" have attached
volumes. These instances cannot be deleted as part of the "batch" data.
Why? Because I said so! No, really this is a bit of an Atmosphere
defined UI business rule. It seems like deleting an instance with
attached volumes in one of the these states (stopped, shelved,
suspended) leads to the Volume not attaching to other instance(s). If
you are a truth seeker (Godspeed, dear reader! GODSPEED) then you may
wish to structure testing experiments to determine this for yourself.

But, when Clouding - it seems like a decent idea to avoid deleting
instances with attached volumes. So - this leads to the idea of *not*
doing that.

Good?

I hope. If not, be upset with me - @lenards