Define edge-case version of Instance Action Impl
Unfortunately, the Atmosphere API does not allow a "terminate" (aka
delete, aka destroy) to be processed by the `run_instance_action`
function defined in service/instance.py. It would seem since there is a
convention in the API _not_ to "delete" but to "end date" - you _could_
make an InstanceActionRequest that would do a "end date" and, thus,
delete. There  may be some _secret_ knowledge I lack for understanding
why a "terminate" is processed as a special case, and via HTTP DELETE.
I wish I knew to ask this of Steve Gregory before his departure.

The original design was to make "terminate" a possible
InstanceActionRequest, like so:
- https://gist.github.com/lenards/85df742edebc596a5b86b646d8c5a145

Some plans don't work and I held onto it for _too_ long.

This function's _specialness_ is hidden within the
instanceActionImpl(..) so that progress in supporting Delete All can be
provided to the community without having to wait for a major
rework/remodel/rethink.

Le mieux est l'ennemi du bien.