Wrap IP addresses in SSH protocol hyperlinks
If we have a username, we include the username in the SSH hyperlinks
to avoid issues with the Host Operating System username differing from
the CyVerse username.

Also, if the session is an emulated one - then don't include that
username - because staff is likely to have the same username, or *at
least* will understand when this doesn't work.