Change ./manage.py maintenance to be easier to test
I was making changes to git_branch, and found that this command would be
affected. I wanted to write some tests to ensure that the command would still
work. However, the previous command would not allow you to pass some arguments and not
others. If you omitted any, it would go into an interactive mode. I couldn't
test whether the defaults were working, because of this mode.

I solved this by adding a flag for --dry-run and removing the interactive
mode.

Other changes:
The start_date format accepts timezones now.
The --start-now flag was removed, this is the default if no --start-date is
specified.
The show command will print a current maintenance if one exists