Allow optimistic updating by cloning the profile object and reverting to it in case of failure
Before doing anything at all, a clone of the profile object is created and
left unmodified. This allows anything to happen to the original profile object
through this file or the external 'dispatch' method. Then, if the Atmo API
call fails, the unmodified cloned profile will be used to revert changes in
the UI. The success notificaiton was enabled since the change in UI is no
longer a guarantee of the change going through to the API.