Fix infinite redirect loop
Atmosphere would redirect the user to login at the authorize endpoint of the
CAS server. After submitting correct creds, CAS would set a cookie so the user
wouldn't have to resubmit creds. The user would be redirected to atmosphere.
Atmosphere would perform its part of the exchange but fail to get a token.
Then it would repeat from the beginning (sending the user to the authorize
endpoint). With the CAS cookie present, the user would be sent right back to
atmosphere as expected, and atmosphere would fail and redirect again.